move FeedStore conformance of core data to its separate file
