+++ EssentialFeed/EssentialFeed/Feed Cache/Infrastructure/CoreData/CoreDataFeedStore.swift
@@ -17,38 +17,6 @@ public final class CoreDataFeedStore: FeedStore {
         context = container.newBackgroundContext()
     }
     
-    public func deleteCachedFeed(completion: @escaping DeletionCompletion) {
-        perform { context in
-            completion(Result {
-                try ManagedCache.find(in: context).map(context.delete).map(context.save)
-            })
-        }
-    }
-    
-    public func insertCachedFeed(_ items: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-        perform { context in
-            completion(Result {
-                let managedCache = try ManagedCache.newUniqueInstance(in: context)
-                managedCache.timestamp = timestamp
-                managedCache.feed = ManagedFeedImage.images(from: items, in: context)
-                
-                try context.save()
-            })
-        }
-    }
-    
-    public func retrievalCachedFeed(completion: @escaping RetrievalCompletion) {
-        perform { context in
-            completion(Result {
-                if let cache = try ManagedCache.find(in: context) {
-                    return (feed: cache.localFeeds, timestamp: cache.timestamp)
-                } else {
-                    return .none
-                }
-            })
-        }
-    }
-    
     func perform(_ action: @escaping (NSManagedObjectContext) -> Void) {
         context.perform { [context] in
             action(context)
